+++ src/plugins/index.js
@@ -1,4 +1,6 @@
 import { nodeInfoPlugin } from './core/nodeInfoPlugin.jsx';
+import { selectionToolsPlugin } from './core/selectionToolsPlugin.jsx';
+import { clipboardPlugin } from './core/clipboardPlugin.jsx';
 import { showcasePlugin } from './examples/showcasePlugin.jsx';
 
-export const corePlugins = [nodeInfoPlugin, showcasePlugin];
+export const corePlugins = [nodeInfoPlugin, selectionToolsPlugin, clipboardPlugin, showcasePlugin];